Downstream of filesystem reorganisation in opm-common
