+++ examples/cpchop_depthtrend.cpp
@@ -27,7 +27,7 @@
 
 #include <opm/common/utility/platform_dependent/reenable_warnings.h>
 
-#include <opm/parser/eclipse/Units/Units.hpp>
+#include <opm/input/eclipse/Units/Units.hpp>
 
 #include <opm/upscaling/CornerpointChopper.hpp>
