+++ examples/mimetic_aniso_solver_test.cpp
@@ -35,8 +35,8 @@
 
 #include "config.h"
 
-#include <opm/parser/eclipse/Parser/Parser.hpp>
-#include <opm/parser/eclipse/Deck/Deck.hpp>
+#include <opm/input/eclipse/Parser/Parser.hpp>
+#include <opm/input/eclipse/Deck/Deck.hpp>
 
 #include <opm/common/utility/platform_dependent/disable_warnings.h>
 
@@ -50,7 +50,7 @@
 
 #include <opm/grid/CpGrid.hpp>
 
-#include <opm/parser/eclipse/Units/Units.hpp>
+#include <opm/input/eclipse/Units/Units.hpp>
 #include <opm/common/utility/parameters/ParameterGroup.hpp>
 
 #include <opm/output/eclipse/EclipseGridInspector.hpp>
