+++ opm/porsol/common/Rock.hpp
@@ -24,7 +24,7 @@
 #include <opm/porsol/common/Matrix.hpp>
 #include <opm/porsol/common/ReservoirPropertyCommon.hpp>
 
-#include <opm/parser/eclipse/Deck/Deck.hpp>
+#include <opm/input/eclipse/Deck/Deck.hpp>
 
 namespace Opm
 {
