+++ opm/porsol/common/SimulatorBase.hpp
@@ -41,7 +41,7 @@
 
 #include <opm/common/utility/numeric/SparseVector.hpp>
 #include <opm/grid/utility/SparseTable.hpp>
-#include <opm/parser/eclipse/Units/Units.hpp>
+#include <opm/input/eclipse/Units/Units.hpp>
 
 #include <opm/grid/common/Volumes.hpp>
 #include <opm/grid/CpGrid.hpp>
