+++ opm/porsol/euler/EulerUpstreamImplicit_impl.hpp
@@ -39,7 +39,7 @@
 
 #include <opm/common/ErrorMacros.hpp>
 #include <opm/core/utility/Average.hpp>
-#include <opm/parser/eclipse/Units/Units.hpp>
+#include <opm/input/eclipse/Units/Units.hpp>
 #include <opm/grid/common/Volumes.hpp>
 #include <opm/grid/utility/StopWatch.hpp>
 #include <opm/porsol/common/ImplicitTransportDefs.hpp>
