+++ opm/porsol/euler/EulerUpstream_impl.hpp
@@ -41,7 +41,7 @@
 
 #include <opm/common/ErrorMacros.hpp>
 #include <opm/core/utility/Average.hpp>
-#include <opm/parser/eclipse/Units/Units.hpp>
+#include <opm/input/eclipse/Units/Units.hpp>
 #include <opm/grid/common/Volumes.hpp>
 #include <opm/porsol/euler/CflCalculator.hpp>
 #include <opm/grid/utility/StopWatch.hpp>
