+++ opm/porsol/euler/ImplicitCapillarity_impl.hpp
@@ -38,7 +38,7 @@
 
 #include <opm/common/ErrorMacros.hpp>
 #include <opm/core/utility/Average.hpp>
-#include <opm/parser/eclipse/Units/Units.hpp>
+#include <opm/input/eclipse/Units/Units.hpp>
 #include <opm/common/utility/numeric/RootFinders.hpp>
 #include <opm/grid/utility/StopWatch.hpp>
 #include <opm/grid/common/Volumes.hpp>
