+++ opm/upscaling/RelPermUtils.hpp
@@ -24,8 +24,8 @@
 #ifndef OPM_UPSCALING_RELPERM_UTILS_HPP
 #define OPM_UPSCALING_RELPERM_UTILS_HPP
 
-#include <opm/parser/eclipse/Parser/Parser.hpp>
-#include <opm/parser/eclipse/Deck/Deck.hpp>
+#include <opm/input/eclipse/Parser/Parser.hpp>
+#include <opm/input/eclipse/Deck/Deck.hpp>
 
 #include <opm/common/utility/numeric/MonotCubicInterpolator.hpp>
