+++ opm/upscaling/initCPGrid.cpp
@@ -32,8 +32,8 @@
 
 #include <opm/common/ErrorMacros.hpp>
 
-#include <opm/parser/eclipse/Parser/Parser.hpp>
-#include <opm/parser/eclipse/Deck/Deck.hpp>
+#include <opm/input/eclipse/Parser/Parser.hpp>
+#include <opm/input/eclipse/Deck/Deck.hpp>
 
 #include <opm/grid/CpGrid.hpp>
