+++ tests/common/test_gravitypressure.cpp
@@ -34,10 +34,10 @@
 #include <opm/upscaling/RelPermUtils.hpp>
 #undef  UNITTEST_TRESPASS_PRIVATE_PROPERTY_DP
 
-#include <opm/parser/eclipse/Parser/Parser.hpp>
-#include <opm/parser/eclipse/Deck/Deck.hpp>
+#include <opm/input/eclipse/Parser/Parser.hpp>
+#include <opm/input/eclipse/Deck/Deck.hpp>
 
-#include <opm/parser/eclipse/Units/Units.hpp>
+#include <opm/input/eclipse/Units/Units.hpp>
 
 #include <cmath>
 #include <cstddef>
